<fix> : 변경된 ddl schema 대응 [지라 이슈 링크]

+++ src/main/java/droni/backend/api/insectcontrol/entity/InsectControlRequest.java
@@ -19,7 +19,7 @@ import java.time.LocalDateTime;
 public class InsectControlRequest {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer insectRequestId;
+    private long insectRequestId;
     @ManyToOne
     @JoinColumn(name = "land_id", nullable = false)
     private Land land;
